CVSL-3983 Removing the old prisoner details updating logic
event processing occurs in the API now
