+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/resource/privateApi/SupportController.kt
@@ -11,11 +11,13 @@ import org.springframework.http.MediaType
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.config.ErrorResponse
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.EligibilityAssessment
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.resource.Tags
+import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.domainEvents.ComAllocatedHandler
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.support.SupportService
 
 @RestController
@@ -23,6 +25,7 @@ import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.support.Sup
 @RequestMapping("/offender", produces = [MediaType.APPLICATION_JSON_VALUE])
 class SupportController(
   private val supportService: SupportService,
+  private val comAllocatedHandler: ComAllocatedHandler,
 ) {
   @GetMapping(
     value = ["/nomisid/{nomsId}/ineligibility-reasons"],
@@ -119,4 +122,38 @@ class SupportController(
   fun getIS91Status(
     @PathVariable nomsId: String,
   ) = supportService.getIS91Status(nomsId)
+
+  @PutMapping(value = ["/sync-com/crn/{crn}"], produces = [MediaType.APPLICATION_JSON_VALUE])
+  @PreAuthorize("hasAnyRole('CVL_ADMIN')")
+  @Operation(
+    summary = "Updates the COM allocation for for licences linked to the given CRN.",
+    description = "Updates the COM allocation for for licences linked to the given CRN.",
+    security = [SecurityRequirement(name = "ROLE_CVL_ADMIN")],
+  )
+  @ApiResponses(
+    value = [
+      ApiResponse(
+        responseCode = "200",
+        description = "The COM allocation info was",
+      ),
+      ApiResponse(
+        responseCode = "401",
+        description = "Unauthorised, requires a valid Oauth2 token",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
+      ),
+      ApiResponse(
+        responseCode = "403",
+        description = "Forbidden, requires an appropriate role",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
+      ),
+      ApiResponse(
+        responseCode = "404",
+        description = "Not-found, an invalid CRN was provided",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
+      ),
+    ],
+  )
+  fun syncComAllocation(
+    @PathVariable crn: String,
+  ) = comAllocatedHandler.syncComAllocation(crn)
 }
