+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/domainEvents/PrisonerUpdatedHandler.kt
@@ -1,12 +1,11 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.domainEvents
 
 import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import tools.jackson.databind.ObjectMapper
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.request.UpdateOffenderDetailsRequest
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.OffenderService
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.conditions.convertToTitleCase
+import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.domainEvents.events.UpdateOffenderDetailsEvent
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.prison.PrisonerSearchApiClient
 
 @Service
@@ -14,18 +13,15 @@ class PrisonerUpdatedHandler(
   private val mapper: ObjectMapper,
   private val offenderService: OffenderService,
   private val prisonerSearchApiClient: PrisonerSearchApiClient,
-  @param:Value("\${update.offender.details.handler.enabled}") private val updateOffenderDetailsHandleEnabled: Boolean,
 ) {
   companion object {
     private val log = LoggerFactory.getLogger(PrisonerUpdatedHandler::class.java)
   }
 
   fun handleEvent(message: String) {
-    if (updateOffenderDetailsHandleEnabled) {
-      val event = mapper.readValue(message, HMPPSPrisonerUpdatedEvent::class.java)
-      if (event.additionalInformation.categoriesChanged.contains(DiffCategory.PERSONAL_DETAILS)) {
-        updatePrisonerDetails(event.additionalInformation.nomsNumber)
-      }
+    val event = mapper.readValue(message, HMPPSPrisonerUpdatedEvent::class.java)
+    if (event.additionalInformation.categoriesChanged.contains(DiffCategory.PERSONAL_DETAILS)) {
+      updatePrisonerDetails(event.additionalInformation.nomsNumber)
     }
   }
 
@@ -35,7 +31,7 @@ class PrisonerUpdatedHandler(
     log.info("processing offender updated event for nomsId: $nomsId")
     offenderService.updateOffenderDetails(
       nomsId,
-      UpdateOffenderDetailsRequest(
+      UpdateOffenderDetailsEvent(
         forename = prisoner.firstName.convertToTitleCase(),
         middleNames = if (prisoner.middleNames == null) "" else prisoner.middleNames.convertToTitleCase(),
         surname = prisoner.lastName.convertToTitleCase(),
