+++ src/test/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/config/LocalStackContainer.kt
@@ -29,7 +29,7 @@ object LocalStackContainer {
     if (localstackIsRunning()) return null
     val logConsumer = Slf4jLogConsumer(log).withPrefix("localstack")
     return LocalStackContainer(
-      DockerImageName.parse("localstack/localstack").withTag("3"),
+      DockerImageName.parse("localstack/localstack").withTag("community-archive"),
     ).apply {
       withServices("sqs", "sns")
       withEnv("DEFAULT_REGION", "eu-west-2")
