+++ src/test/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/integration/domainEvents/DomainEventsListenerIntegrationTest.kt
@@ -38,7 +38,7 @@ import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.domainEvent
 import java.time.Duration
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-@TestPropertySource(properties = ["domain.event.listener.disabled=false", "update.offender.details.handler.enabled=true"])
+@TestPropertySource(properties = ["domain.event.listener.disabled=false"])
 class DomainEventsListenerIntegrationTest : IntegrationTestBase() {
 
   @MockitoSpyBean
