--- src/test/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/resource/privateApi/OffenderControllerTest.kt
@@ -1,76 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.resource.privateApi
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.mockito.kotlin.reset
-import org.mockito.kotlin.times
-import org.mockito.kotlin.verify
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.http.MediaType
-import org.springframework.test.context.bean.override.mockito.MockitoBean
-import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.config.ControllerAdvice
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.config.NotSecuredWebMvcTest
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model.request.UpdateOffenderDetailsRequest
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.OffenderService
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.StaffService
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.domainEvents.ComAllocatedHandler
-import java.time.LocalDate
-
-@NotSecuredWebMvcTest(controllers = [OffenderController::class])
-class OffenderControllerTest {
-
-  @MockitoBean
-  private lateinit var comAllocatedHandler: ComAllocatedHandler
-
-  @MockitoBean
-  private lateinit var offenderService: OffenderService
-
-  @MockitoBean
-  private lateinit var staffService: StaffService
-
-  @Autowired
-  private lateinit var mvc: MockMvc
-
-  @Autowired
-  private lateinit var mapper: ObjectMapper
-
-  @BeforeEach
-  fun reset() {
-    reset(offenderService, staffService)
-
-    mvc = MockMvcBuilders
-      .standaloneSetup(
-        OffenderController(
-          comAllocatedHandler,
-          offenderService,
-          updateOffenderDetailsHandleEnabled = false,
-        ),
-      )
-      .setControllerAdvice(ControllerAdvice())
-      .build()
-  }
-
-  @Test
-  fun `update offender details`() {
-    val body = UpdateOffenderDetailsRequest(
-      forename = "John",
-      middleNames = "Joe Jim",
-      surname = "Bloggs-Doe",
-      dateOfBirth = LocalDate.parse("1970-01-01"),
-    )
-
-    val request = put("/offender/nomisid/exampleNomisId/update-offender-details")
-      .accept(MediaType.APPLICATION_JSON)
-      .contentType(MediaType.APPLICATION_JSON)
-      .content(mapper.writeValueAsBytes(body))
-
-    mvc.perform(request).andExpect(status().isOk)
-
-    verify(offenderService, times(1)).updateOffenderDetails("exampleNomisId", body)
-  }
-}
